Remove Express server and complete migration to serverless Lambda architecture
- Created new adminHandler.ts Lambda function for OAuth and admin endpoints
- Removed Express server and all related dependencies
- Updated Terraform to deploy separate admin Lambda with its own API Gateway
- Updated frontend to use CloudFront routes for all admin endpoints
- Removed backend Docker service from docker-compose.yml
- Updated documentation to reflect fully serverless architecture

Benefits:
- Eliminated always-on server costs (pay-per-request only)
- Improved scalability with automatic Lambda scaling
- Simplified deployment with single architecture for all environments
- Better security isolation between public and admin functions
