--- backend/Dockerfile.dev
@@ -1,21 +1,0 @@
-FROM node:18-alpine
-
-WORKDIR /app
-
-# Copy package files
-COPY package*.json ./
-
-# Install dependencies
-RUN npm install
-
-# Copy source code
-COPY . .
-
-# Create dist directory
-RUN mkdir -p dist
-
-# Expose port
-EXPOSE 3001
-
-# Start development server
-CMD ["npm", "run", "dev"]
