+++ backend/package.json
@@ -4,10 +4,9 @@
   "description": "Lambda functions for Chautauqua calendar generation",
   "main": "index.js",
   "scripts": {
-    "build": "npm run test && npm run build:dev",
-    "build:dev": "npx esbuild src/handlers/*.ts --bundle --platform=node --target=node24 --outdir=dist --external:aws-sdk && npx esbuild src/services/*.ts --bundle --platform=node --target=node18 --outdir=dist/services --format=cjs --external:aws-sdk --external:@aws-sdk/client-dynamodb --external:@aws-sdk/lib-dynamodb --external:axios --external:cheerio --external:date-fns && npx esbuild src/server.ts --bundle --platform=node --target=node18 --outfile=dist/server.js --format=cjs --external:aws-sdk",
-    "build:prod": "npx esbuild src/handlers/calendarHandler.ts --bundle --platform=node --target=node24 --outfile=dist/calendarHandler.js --external:aws-sdk --external:@aws-sdk/client-dynamodb --external:@aws-sdk/lib-dynamodb --external:ical-generator --external:date-fns --external:cheerio --external:axios --external:uuid --external:node-fetch",
-    "dev": "node dist/server.js",
+    "build": "npm run test && npm run build:prod",
+    "build:prod": "npx esbuild src/handlers/calendarHandler.ts --bundle --platform=node --target=node24 --outfile=dist/calendarHandler.js --external:aws-sdk --external:@aws-sdk/client-dynamodb --external:@aws-sdk/lib-dynamodb --external:ical-generator --external:date-fns --external:cheerio --external:axios --external:uuid --external:node-fetch && npx esbuild src/handlers/adminHandler.ts --bundle --platform=node --target=node24 --outfile=dist/adminHandler.js --external:aws-sdk --external:@aws-sdk/client-dynamodb --external:@aws-sdk/lib-dynamodb --external:googleapis --external:jsonwebtoken --external:node-fetch --external:uuid",
+    "build:services": "npx esbuild src/services/*.ts --bundle --platform=node --target=node18 --outdir=dist/services --format=cjs --external:aws-sdk --external:@aws-sdk/client-dynamodb --external:@aws-sdk/lib-dynamodb --external:axios --external:cheerio --external:date-fns",
     "deploy": "npm run build && zip -r lambda-function.zip dist/ package.json && aws lambda update-function-code --function-name chautauqua-calendar-generator --zip-file fileb://lambda-function.zip",
     "package": "npm run build:prod && zip -r lambda-function.zip dist/ package.json node_modules/",
     "test": "jest --passWithNoTests",
@@ -26,9 +25,7 @@
     "aws-lambda": "^1.0.7",
     "axios": "^1.6.0",
     "cheerio": "^1.0.0-rc.12",
-    "cors": "^2.8.5",
     "date-fns": "^2.29.3",
-    "express": "^4.18.2",
     "googleapis": "^153.0.0",
     "ical-generator": "^4.1.0",
     "jsonwebtoken": "^9.0.2",
@@ -40,8 +37,6 @@
   "devDependencies": {
     "@types/aws-lambda": "^8.10.109",
     "@types/cheerio": "^0.22.31",
-    "@types/cors": "^2.8.13",
-    "@types/express": "^4.17.17",
     "@types/jest": "^30.0.0",
     "@types/jsonwebtoken": "^9.0.10",
     "@types/node": "^24.0.0",
